[fix] 파일위치 변경 및 import수정
