+++ src/components/layout/Gnb/Logo.tsx
@@ -1,5 +1,5 @@
-import Image from "next/image";
-import Link from "next/link";
+import Image from 'next/image';
+import Link from 'next/link';
 
 export default function Logo() {
   return (
